Convert ServiceUploadDetail.js to use ES6 modules; add complete PropTypes
