+++ spec/javascripts/service_uploads/PastServiceUploads.test.js
@@ -21,6 +21,7 @@ describe('with valid past service upload data', function () {
       <PastServiceUploads
         serviceUploads={[
           {
+            id: 3,
             file_name: 'test_file.csv',
             created_at: '2018-01-03T14:38:38.505Z', // Rails created_at
             services: [],
