+++ ui/legacy.js
@@ -36,6 +36,5 @@ import '../app/assets/javascripts/student_profile/page_container.jsx';
 
 // all the rest, unordered
 import '../app/assets/javascripts/service_uploads/new_service_upload.jsx';
-import '../app/assets/javascripts/service_uploads/service_upload_detail.jsx';
 import '../app/assets/javascripts/service_uploads/service_uploads_page.jsx';
